Add files via upload
Market efficiency occurs at the expense of acknowledging your customer’s feedback and obtaining their 
trust. In particular, the global insurance market, having to bear the consequences of the covid-19 outbreak in 
the last year, faces several challenges regarding the realization and confrontation of customers' main 
concerns. With a few numbers of literature addressing this matter, there is a crucial need for more exploration 
in the district of customers’ sentiments. This study aims to utilize social media as an impartial platform to 
obtain a clear understanding of the efficiency of Canadian health insurance during the vaccination months of 
the pandemic. In this study, 1000 Twitter posts containing the keywords of “health insurance” and “Canada” 
are collected as the database. The Microblog sentiment analysis is performed, adopting two lexicon-based 
models of Textblob module and NRC emotion lexicon to indicate the polarity, subjectivity, and emotions of 
the customers toward the health insurance of Canada within the 1st of December 2020 until the 28th of 
February 2021. The results suggest an overall positive polarity toward the health insurance of Canada with 
the subjectivity of mostly factual information. The fear of bankruptcy and heavy taxation is indicated to be 
the biggest despair during the vaccination months. However, the trust of the government and medical 
providers including the specialists is strongly present. The contrast between the fear and sadness associated 
with bankruptcy and the strong positive emotion towards the word “job” concludes the fear to be mostly 
present among the self-employed due to their lack of public insurance. The foreign health care plans are also 
determined to be inefficient.